+++ spackle-android-lib/src/main/kotlin/co/electriccoin/zcash/spackle/BundleCompat.kt
@@ -1,12 +1,11 @@
 package co.electriccoin.zcash.spackle
 
-import android.os.Build
 import android.os.Bundle
 import java.io.Serializable
 
 @Suppress("DEPRECATION")
 inline fun <reified T : Serializable> Bundle.getSerializableCompat(key: String): T? =
-    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+    if (AndroidApiVersion.isAtLeastTiramisu) {
         getSerializable(key, T::class.java)
     } else {
         getSerializable(key) as? T