+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/account/fixture/TransactionUiFixture.kt
@@ -6,6 +6,7 @@ import cash.z.ecc.android.sdk.model.TransactionOverview
 import cash.z.ecc.android.sdk.model.TransactionRecipient
 import cash.z.ecc.android.sdk.model.ZcashNetwork
 import cash.z.ecc.android.sdk.type.AddressType
+import co.electriccoin.zcash.ui.common.model.AddressBookContact
 import co.electriccoin.zcash.ui.screen.account.model.TransactionUi
 import co.electriccoin.zcash.ui.screen.account.model.TrxItemState
 
@@ -23,18 +24,22 @@ object TransactionUiFixture {
 
     val MESSAGES: List<String> = listOf("Thanks for the coffee", "It was great to meet you!")
 
+    val ADDRESS_BOOK_CONTACT: AddressBookContact? = null
+
+    @Suppress("LongParameterList")
     internal fun new(
         overview: TransactionOverview = OVERVIEW,
         recipient: TransactionRecipient = RECIPIENT,
         recipientAddressType: AddressType = RECIPIENT_ADDRESS_TYPE,
         expandableState: TrxItemState = EXPANDABLE_STATE,
         messages: List<String> = MESSAGES,
+        addressBookContact: AddressBookContact? = ADDRESS_BOOK_CONTACT
     ) = TransactionUi(
         overview = overview,
         recipient = recipient,
         recipientAddressType = recipientAddressType,
         expandableState = expandableState,
         messages = messages,
-        addressBookContact = null
+        addressBookContact = addressBookContact
     )
 }