+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/qrcode/util/AndroidQrCodeImageGenerator.kt
@@ -1,4 +1,4 @@
-package co.electriccoin.zcash.ui.screen.receive.util
+package co.electriccoin.zcash.ui.screen.qrcode.util
 
 import android.graphics.Bitmap
 import android.graphics.Color